Allow boilerplate body text within updated issue
If someone has changed the boilerplate title, but left in the
boilerplate body, adding their issue below it, then their issue will be
closed even though it may be a valid issue. With this commit an issue
whose body ends with the boilerplate text (including any trailing
whitespace or newlines to avoid someone hitting enter after the
boilerplate constituting an issue) will be closed, but an issue
that contains text after the boilerplate text will not.

Fixes #4